+++ web/app/components/workflow/nodes/if-else/default.ts
@@ -1,7 +1,6 @@
 import { BlockEnum, type NodeDefault } from '../../types'
 import { type IfElseNodeType, LogicalOperator } from './types'
 import { isEmptyRelatedOperator } from './utils'
-import { TransferMethod } from '@/types/app'
 import { ALL_CHAT_AVAILABLE_BLOCKS, ALL_COMPLETION_AVAILABLE_BLOCKS } from '@/app/components/workflow/blocks'
 const i18nPrefix = 'workflow.errorMsg'
 
@@ -79,18 +78,3 @@ const nodeDefault: NodeDefault<IfElseNodeType> = {
 }
 
 export default nodeDefault
-
-export const FILE_TYPE_OPTIONS = [
-  { value: 'image', i18nKey: 'image' },
-  { value: 'document', i18nKey: 'doc' },
-  { value: 'audio', i18nKey: 'audio' },
-  { value: 'video', i18nKey: 'video' },
-]
-
-export const TRANSFER_METHOD = [
-  { value: TransferMethod.local_file, i18nKey: 'localUpload' },
-  { value: TransferMethod.remote_url, i18nKey: 'url' },
-]
-
-export const SUB_VARIABLES = ['type', 'size', 'name', 'url', 'extension', 'mime_type', 'transfer_method']
-export const OUTPUT_FILE_SUB_VARIABLES = SUB_VARIABLES.filter(key => key !== 'transfer_method')
